Refactored spdump structs
