+++ gazelle/internal/spdump/BUILD.bazel
@@ -13,7 +13,14 @@ filegroup(
 
 go_library(
     name = "spdump",
-    srcs = ["manifest.go"],
+    srcs = [
+        "dependency.go",
+        "manifest.go",
+        "platform.go",
+        "product.go",
+        "target.go",
+        "target_dependency.go",
+    ],
     importpath = "github.com/cgrindel/swift_bazel/gazelle/internal/spdump",
     visibility = ["//gazelle:__subpackages__"],
     deps = [
@@ -24,7 +31,11 @@ go_library(
 
 go_test(
     name = "spdump_test",
-    srcs = ["manifest_test.go"],
+    srcs = [
+        "manifest_test.go",
+        "target_dependency_test.go",
+        "target_test.go",
+    ],
     deps = [
         ":spdump",
         "@com_github_stretchr_testify//assert",
