+++ gazelle/internal/spdump/platform.go
@@ -1,0 +1,7 @@
+package spdump
+
+type Platform struct {
+	Name    string `json:"platformName"`
+	Version string
+}
+
