+++ gazelle/internal/spdump/target_dependency_test.go
@@ -1,0 +1,28 @@
+package spdump_test
+
+import (
+	"testing"
+
+	"github.com/cgrindel/swift_bazel/gazelle/internal/spdump"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestTargetDependencyImportName(t *testing.T) {
+	t.Run("product", func(t *testing.T) {
+		td := spdump.TargetDependency{
+			Product: &spdump.ProductReference{
+				ProductName:    "ArgumentParser",
+				DependencyName: "swift-argument-parser",
+			},
+		}
+		actual := td.ImportName()
+		assert.Equal(t, "ArgumentParser", actual)
+	})
+	t.Run("by name", func(t *testing.T) {
+		td := spdump.TargetDependency{
+			ByName: &spdump.ByNameReference{TargetName: "MySwiftPackage"},
+		}
+		actual := td.ImportName()
+		assert.Equal(t, "MySwiftPackage", actual)
+	})
+}
